% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_multiple.R
\name{broom_multiple}
\alias{broom_multiple}
\title{Function to apply all of \strong{broom}'s tidying functions and return a list.}
\usage{
broom_multiple(model, transformed = FALSE)
}
\arguments{
\item{model}{A model object.}

\item{transformed}{Should the transformed original data be returned too? 
Default is \code{FALSE}.}
}
\value{
List.
}
\description{
Function to apply all of \strong{broom}'s tidying functions and return a list.
}
\author{
Stuart K. Grange
}
