% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_helpers.R
\name{broom_sweep}
\alias{broom_sweep}
\alias{extract_intercept}
\alias{extract_slope}
\title{Function to further clean \strong{broom}'s tidy data frames.}
\usage{
broom_sweep(fit, extra = TRUE, round = 6)

extract_slope(df)

extract_intercept(df)
}
\description{
Function to further clean \strong{broom}'s tidy data frames.
}
\author{
Stuart K. Grange
}

