% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposing_functions.R
\name{decompose_kz}
\alias{decompose_kz}
\title{Function to decompose time-series with Kolmogorov-Zurbenko filters.}
\usage{
decompose_kz(df, variable = "value", plot = TRUE, invalidate = TRUE)
}
\description{
Function to decompose time-series with Kolmogorov-Zurbenko filters.
}
\author{
Stuart K. Grange
}
