% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposing_functions.R
\name{decompose_stl}
\alias{decompose_stl}
\title{Function to decompose time-series with loess.}
\usage{
decompose_stl(df, variable = "value", plot = FALSE, invalidate = TRUE)
}
\description{
Function to decompose time-series with loess.
}
\author{
Stuart K. Grange
}
