% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_helpers.R
\name{ggimportance}
\alias{ggimportance}
\title{Function to plot the \code{increase_in_mse} variable for a \code{randomForest} 
object.}
\usage{
ggimportance(list_model)
}
\arguments{
\item{list_model}{A \code{randomForest} object.}
}
\value{
Invisible.
}
\description{
Function to plot the \code{increase_in_mse} variable for a \code{randomForest} 
object.
}
\seealso{
\code{\link{randomForest}}, \code{\link{importance_tidy}}
}
\author{
Stuart K. Grange
}
