% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_helpers.R
\name{ggpartial}
\alias{ggpartial}
\title{Function to plot partial dependence plots for a \code{randomForest} object.}
\usage{
ggpartial(list_model, df, variable)
}
\arguments{
\item{list_model}{A \code{randomForest} object.}

\item{df}{Data frame to predict.}

\item{variable}{Variable names to predict.}
}
\value{
Invisible.
}
\description{
Function to plot partial dependence plots for a \code{randomForest} object.
}
\seealso{
\code{\link{randomForest}}, \code{\link{importance_tidy}}, 
\code{\link{ggimportance}}
}
\author{
Stuart K. Grange
}
