% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_helpers.R
\name{ggtrees}
\alias{ggtrees}
\title{Function to plot the MSE and number of trees for a \code{randomForest} object.}
\usage{
ggtrees(list_model)
}
\arguments{
\item{list_model}{A \code{randomForest} object.}
}
\value{
Invisible.
}
\description{
Function to plot the MSE and number of trees for a \code{randomForest} object.
}
\seealso{
\code{\link{randomForest}}
}
\author{
Stuart K. Grange
}
