% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_helpers.R
\name{importance_tidy}
\alias{importance_tidy}
\title{Function to tidy the \code{importance} matrix for a \code{randomForest} object.}
\usage{
importance_tidy(list_model, scale = TRUE)
}
\arguments{
\item{list_model}{A \code{randomForest} object.}

\item{scale}{Should the importance variables be scaled?}
}
\value{
Data frame.
}
\description{
Function to tidy the \code{importance} matrix for a \code{randomForest} object.
}
\seealso{
\code{\link{randomForest}}, \code{\link{ggimportance}}
}
\author{
Stuart K. Grange
}
