% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_surface.R
\name{interpolate_surface}
\alias{interpolate_surface}
\title{Function to conveniently interpolate a reglar surface from irregular 
observations.}
\usage{
interpolate_surface(df, x = "x", y = "y", z = "z", smooth = TRUE,
  length = 100)
}
\arguments{
\item{df}{Data frame.}

\item{x}{Variable.}

\item{y}{Variable.}

\item{z}{Variable.}

\item{smooth}{Should the surface be smoothed?}

\item{length}{Length of grid to interpolate to.}
}
\value{
Data frame.
}
\description{
\code{interpolate_surface} uses the \strong{akima} package to interpolate
and the \strong{fields} package to smooth the surface.
}
\author{
Stuart K. Grange
}
