% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_smoother.R
\name{kernel_smoother}
\alias{kernel_smoother}
\title{Function to apply kernel smoothing functions.}
\usage{
kernel_smoother(x, kernel = "gaussian")
}
\arguments{
\item{x}{Vector to transform/smooth.}

\item{kernel}{Type of smoothing kernel to apply to \code{x}. See details 
for supported methods.}
}
\value{
Vector the length of \code{x}.
}
\description{
\code{kernel_smoother} is used to transform a vector.
}
\details{
Current kernels which are supported are:
 
\itemize{
 \item Gaussian/normal
 \item Epanechnikov
 \item Logistic
 \item Cosine
 \item Triangular
 \item Box/uniform
 \item Tricube
 \item Triweight
 \item Biweight/quartic
}
}
\examples{

# Create vector
vector <- seq(-5, 5, by = 0.1)

# Transform
vector_smooth <- kernel_smoother(vector, "gaussian")

# Or use a different kernel
vector_smooth_epanechnikov <- kernel_smoother(vector, "epanechnikov")
# plot(vector_smooth_epanechnikov, type = "l")

}
\seealso{
\href{https://en.wikipedia.org/wiki/Kernel_(statistics)}{wikipedia}
}
\author{
Stuart K. Grange
}
