% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_regression_functions.R
\name{least_squares_regression}
\alias{least_squares_regression}
\title{Function to apply simple linear regression techniques to observations.}
\usage{
least_squares_regression(df, x, y, zero_intercept = FALSE, weights = NULL,
  round = 6, plot = FALSE)
}
\description{
Function to apply simple linear regression techniques to observations.
}
\author{
Stuart K. Grange
}

