% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_map.R
\name{linear_map}
\alias{linear_map}
\title{Function to apply a linear map transformation to a numeric vector.}
\usage{
linear_map(x, from, to)
}
\arguments{
\item{x}{Numeric vector.}

\item{from}{Starting value for linear map.}

\item{to}{End value for linear map.}
}
\description{
Function to apply a linear map transformation to a numeric vector.
}
\author{
Stuart K. Grange
}
