% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prediction.R
\name{make_prediction}
\alias{make_prediction}
\title{Function to predict using a model object and return a vector.}
\usage{
make_prediction(model, df)
}
\arguments{
\item{model}{Model object which can be used with \code{\link{predict}}.}

\item{df}{Data frame containing input data which can be used by \code{model},
not always needed.}
}
\value{
Numeric vector.
}
\description{
Function to predict using a model object and return a vector.
}
\examples{
\dontrun{

# Predict using a random forest model
value_predict <- make_prediction(list_random_forest, data_testing_set)

}

}
\author{
Stuart K. Grange
}
