% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_quantiles.R
\name{many_quantiles}
\alias{many_quantiles}
\title{Function to apply many quantile regression models and return a tidy data 
frame.}
\usage{
many_quantiles(df, y, x, taus = seq(0.05, 0.95, by = 0.05), round = 6)
}
\arguments{
\item{df}{Data frame.}

\item{y}{Variable which is dependent on \code{x}.}

\item{x}{Variable which predicts \code{y}.}

\item{taus}{Quantiles to plot. Default uses \code{seq(0.05, 0.95, by = 0.05)}.}

\item{round}{Decimal point to round all numeric variables in the returned 
data frame. Default is 6.}
}
\value{
Data frame with model coefficents.
}
\description{
Function to apply many quantile regression models and return a tidy data 
frame.
}
\author{
Stuart K. Grange
}
