% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorological_functions.R
\name{normalise_for_meteorology}
\alias{normalise_for_meteorology}
\title{Function to normalise a concentration variable based on "average" 
meteorological conditions.}
\usage{
normalise_for_meteorology(list_model, df, variables = NA, n = 100)
}
\value{
Data frame.
}
\description{
Function to normalise a concentration variable based on "average" 
meteorological conditions.
}
\author{
Stuart K. Grange
}
