% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_many_quantiles.R
\name{plot_many_quantiles}
\alias{plot_many_quantiles}
\title{Function to quickly plot many quantiles regression lines on a scatterplot.}
\usage{
plot_many_quantiles(df, y, x, type = "scatterplot", taus = seq(0.05, 0.95,
  by = 0.05), square = FALSE)
}
\arguments{
\item{df}{Data frame.}

\item{y}{Variable which is dependent on \code{x}.}

\item{x}{Variable which predicts \code{y}.}

\item{type}{Type of plot. Can either be \code{"scatterplot"} or 
\code{"coefficients"}.}

\item{taus}{Quantiles to plot. Default uses \code{seq(0.05, 0.95, by = 0.05)}.}

\item{square}{Should the data be plotted with square axes? Default is 
\code{FALSE}.}
}
\description{
Function to quickly plot many quantiles regression lines on a scatterplot.
}
\author{
Stuart K. Grange
}
