% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorological_functions.R
\name{prepare_input_data}
\alias{prepare_input_data}
\title{Function to prepare meteorological and air quality data for modelling.}
\usage{
prepare_input_data(df)
}
\value{
A named list containing two data frames.
}
\description{
Function to prepare meteorological and air quality data for modelling.
}
\author{
Stuart K. Grange
}
