% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_model_coefficients.R
\name{pull_model_coefficients}
\alias{pull_model_coefficients}
\alias{pull_model_r_squared}
\alias{pull_model_mse}
\alias{pull_model_slope}
\alias{pull_model_intercept}
\alias{pull_model_p_value}
\alias{pull_model_f_statistic}
\alias{pull_model_degrees_of_freedom}
\title{Functions to pull useful pieces out of model objects.}
\usage{
pull_model_coefficients(model, named = FALSE)

pull_model_r_squared(model)

pull_model_mse(model)

pull_model_slope(model)

pull_model_intercept(model)

pull_model_p_value(model)

pull_model_f_statistic(model)

pull_model_degrees_of_freedom(model)
}
\arguments{
\item{model}{A model object.}

\item{named}{Should the return vector be named?}
}
\value{
A vector, either named or unnamed.
}
\description{
Functions to pull useful pieces out of model objects.
}
\author{
Stuart K. Grange
}
