% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rows.R
\name{random_rows}
\alias{random_rows}
\title{Function to get a vector of random rows for data frame sampling, usually for
creating training and testing data frames for models.}
\usage{
random_rows(df, seed = 123, fraction = 0.8)
}
\arguments{
\item{df}{Data frame to get random rows from.}

\item{seed}{What \code{set.seed} should be used?}

\item{fraction}{Fraction of \code{df} to sample.}
}
\value{
Integer vector.
}
\description{
Function to get a vector of random rows for data frame sampling, usually for
creating training and testing data frames for models.
}
\author{
Stuart K. Grange
}
