% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_parametric_linear_regression_functions.R
\name{theil_sen_estimator}
\alias{siegel_repeated_medians}
\alias{theil_sen_estimator}
\title{Functions to apply non-parametric linear regression methods to observations.}
\usage{
theil_sen_estimator(df, x, y, round = 6, plot = FALSE)

siegel_repeated_medians(df, x, y, round = 6, plot = FALSE)
}
\description{
The Theil-Sen single median and Siegel repeated medians regression techniques 
are considered robust regression techniques and often suitable when parametric
techniques fail or their assumptions are violated.
}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{mblm}}
}

