% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_parametric_linear_regression_functions.R
\name{theil_sen_prewhiten}
\alias{theil_sen_prewhiten}
\title{Function to calculate the Thiel-Sen estimate of slope after the application 
of a prewhitening algorithm.}
\usage{
theil_sen_prewhiten(x, method = "yuepilon", round = 6, plot = FALSE)
}
\arguments{
\item{x}{Vector of data. This will usually be a padded timeseries.}

\item{method}{Prewhitening method to use. Can be \code{"yuepilon"} or 
\code{"zhang"} and default is \code{"yuepilon"}.}

\item{round}{Returned coefficients precision. Default is 6 dp.}
}
\value{
Data frame of coefficients.
}
\description{
The prewhitening algorithm helps deal with autocorrelation.
}
\author{
Stuart K. Grange
}
