% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{least_squares_regression}
\alias{least_squares_regression}
\alias{quantile_regression}
\alias{robust_linear_regression}
\title{Function to apply linear regression techniques to observations.}
\usage{
least_squares_regression(df, y, x, zero_intercept = FALSE, weights = NULL,
  round = 6, plot = FALSE)

robust_linear_regression(df, y, x, zero_intercept = FALSE, weights = NULL,
  method = "M", round = 6, plot = FALSE)

quantile_regression(df, y, x, zero_intercept = FALSE, weights = NULL,
  tau = 0.5, method = "br", round = 6, plot = FALSE)
}
\description{
\code{y} is a function of \code{x} or \code{y = y(x)}.
}
\author{
Stuart K. Grange
}

